fix some weird storybook build issues
