+++ packages/dialtone-vue/.storybook/preview.js
@@ -14,7 +14,7 @@ import fixDefaultSlot from '../components/plugins/fixDefaultSlot';
 import { setEmojiAssetUrlSmall, setEmojiAssetUrlLarge, setCustomEmojiUrl, setCustomEmojiJson } from '@/common/emoji';
 import customEmojiJson from '@/common/custom-emoji.json';
 import { dialtoneDarkTheme, dialtoneLightTheme } from './dialtone-themes.js';
-import { DialtoneDocsPage } from './DialtoneDocsPage.jsx';
+import { DialtoneDocsPage } from './DialtoneDocsPage.js';
 import { DtTooltipDirective } from '@/directives/tooltip_directive';
 import { DtScrollbarDirective } from '@/directives/scrollbar_directive';
 import { DtStack } from '@/components/stack';
@@ -121,13 +121,10 @@ export default {
           };
         }, [channel, setDark]);
 
-        return (
-          <DocsContainer
-            theme={isDark ? dialtoneDarkTheme : dialtoneLightTheme}
-            context={props.context}
-          >
-            {children}
-          </DocsContainer>
+        return React.createElement(
+          DocsContainer,
+          { theme: isDark ? dialtoneDarkTheme : dialtoneLightTheme, context: props.context },
+          children,
         );
       },
     },
@@ -135,5 +132,5 @@ export default {
     percy: { globalShow: true },
   },
 
-  tags: ['autodocs']
+  tags: ['autodocs'],
 };
